Remove anss.util and anss.edgethread packages
Removed all dependencies for anss.util and edgethread packages. Most of these packages were unused edge server related code, sql handling code, and various util methods that were not needed.

SeedUtil.java was still needed in part, it was stripped down, and moved into anss.cd11.

A time formatting method was not able to be removed, but was instead updated to use the SimpleDateFormat class and placed in ChannelSubframe directly.

Cleaned up the cd11 package. It had numerous warning and dead code.